main refactored does it work?
